Address CodeRabbit review: abort checks, batcher flushes, emoji streaming, test cleanup
- Add abortSignal checks in path resolver glob/file reads after awaits
- Flush StreamBatcher before tool-call/tool-status/tool-result/usage events
- Switch text-stream filtering to filterStreamChunk (thoughts use filterText)
  and drain emojiFilter.flushBuffer() in flush()
- Extract shared test scaffolding into zed-test-helpers.ts (DRY)
- Add afterEach session disposal in prompt/stale tests (listener leak fix)
- Add test for string-typed line/offset arg coercion
