+++ packages/cli/src/zed-integration/zedIntegration.prompt.test.ts
@@ -4,238 +4,32 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { describe, expect, it, vi } from 'vitest';
+import { afterEach, describe, expect, it } from 'vitest';
 import type * as acp from '@agentclientprotocol/sdk';
-import type { AgentEvent, Agent } from '@vybestack/llxprt-code-agents';
-import {
-  ToolConfirmationOutcome,
-  type ToolConfirmationPayload,
-} from '@vybestack/llxprt-code-tools';
-import type { ApprovalMode, Config } from '@vybestack/llxprt-code-core';
+import { ToolConfirmationOutcome } from '@vybestack/llxprt-code-tools';
+import type { Config } from '@vybestack/llxprt-code-core';
 import { todoEvents } from '@vybestack/llxprt-code-core';
 
 import { Session } from './zedIntegration.js';
+import {
+  buildFakeAgent,
+  buildScriptedAgent,
+  RecordingConnection,
+  buildMinimalConfig,
+  createSession,
+  runPrompt,
+  editConfirmation,
+} from './zed-test-helpers.js';
 
-type ConfirmationCapture = {
-  confirmationId: string;
-  decision: ToolConfirmationOutcome;
-  payload?: ToolConfirmationPayload;
-  requiresUserConfirmation?: boolean;
-};
-
-function buildFakeAgent(events: readonly AgentEvent[]): {
-  agent: Agent;
-  confirmations: ConfirmationCapture[];
-} {
-  return buildScriptedAgent(() => events);
-}
-
-function buildScriptedAgent(nextEvents: () => readonly AgentEvent[]): {
-  agent: Agent;
-  confirmations: ConfirmationCapture[];
-} {
-  const confirmations: ConfirmationCapture[] = [];
-  const agent = {
-    async *stream(_input: unknown, _opts?: unknown): AsyncIterable<AgentEvent> {
-      for (const e of nextEvents()) {
-        yield e;
-      }
-    },
-    getApprovalMode: (): ApprovalMode => 'default' as ApprovalMode,
-    setApprovalMode: vi.fn(),
-    dispose: vi.fn().mockResolvedValue(undefined),
-    tools: {
-      respondToConfirmation: (
-        confirmationId: string,
-        decision: ToolConfirmationOutcome,
-        payload?: ToolConfirmationPayload,
-        requiresUserConfirmation?: boolean,
-      ) => {
-        confirmations.push({
-          confirmationId,
-          decision,
-          ...(payload === undefined ? {} : { payload }),
-          ...(requiresUserConfirmation === undefined
-            ? {}
-            : { requiresUserConfirmation }),
-        });
-      },
-      onConfirmationRequest: () => () => {},
-      onToolUpdate: () => () => {},
-      setEditorCallbacks: () => {},
-      setEnabled: vi.fn().mockResolvedValue(undefined),
-      list: () => [],
-      keys: {},
-    },
-  } as unknown as Agent;
-  return { agent, confirmations };
-}
-
-class RecordingConnection {
-  readonly messages: Array<
-    | { kind: 'sessionUpdate'; update: acp.SessionUpdate }
-    | {
-        kind: 'requestPermission';
-        request: acp.RequestPermissionRequest;
-        outcome: acp.RequestPermissionOutcome;
-      }
-  > = [];
-  private permissionOutcome: acp.RequestPermissionOutcome = {
-    outcome: 'selected',
-    optionId: ToolConfirmationOutcome.ProceedOnce,
-  };
-  private permissionRejection: Error | null = null;
-  private gatedDeferred: {
-    resolve: (o: acp.RequestPermissionOutcome) => void;
-    promise: Promise<acp.RequestPermissionOutcome>;
-  } | null = null;
-  private gatedArrived: (() => void) | null = null;
-
-  setPermissionOutcome(outcome: acp.RequestPermissionOutcome): void {
-    this.permissionOutcome = outcome;
-  }
-
-  rejectPermission(error: Error): void {
-    this.permissionRejection = error;
-  }
-
-  armPermissionGate(): {
-    arrived: Promise<void>;
-    settle: (o: acp.RequestPermissionOutcome) => void;
-  } {
-    let resolveArrived!: () => void;
-    const arrived = new Promise<void>((r) => {
-      resolveArrived = r;
-    });
-    let resolvePermission!: (o: acp.RequestPermissionOutcome) => void;
-    const promise = new Promise<acp.RequestPermissionOutcome>((r) => {
-      resolvePermission = r;
-    });
-    this.gatedDeferred = { resolve: resolvePermission, promise };
-    this.gatedArrived = resolveArrived;
-    return {
-      arrived,
-      settle: (o: acp.RequestPermissionOutcome) => {
-        const d = this.gatedDeferred;
-        this.gatedDeferred = null;
-        this.gatedArrived = null;
-        d?.resolve(o);
-      },
-    };
-  }
-
-  sessionUpdate = vi.fn(
-    async (params: acp.SessionNotification): Promise<void> => {
-      this.messages.push({ kind: 'sessionUpdate', update: params.update });
-    },
-  );
-
-  requestPermission = vi.fn(
-    async (
-      params: acp.RequestPermissionRequest,
-    ): Promise<acp.RequestPermissionResponse> => {
-      if (this.permissionRejection !== null) {
-        throw this.permissionRejection;
-      }
-      if (this.gatedDeferred !== null) {
-        this.messages.push({
-          kind: 'requestPermission',
-          request: params,
-          outcome: {
-            outcome: 'selected',
-            optionId: ToolConfirmationOutcome.Cancel,
-          },
-        });
-        const arrivedFn = this.gatedArrived;
-        this.gatedArrived = null;
-        arrivedFn?.();
-        const outcome = await this.gatedDeferred.promise;
-        this.gatedDeferred = null;
-        return { outcome };
-      }
-      this.messages.push({
-        kind: 'requestPermission',
-        request: params,
-        outcome: this.permissionOutcome,
-      });
-      return { outcome: this.permissionOutcome };
-    },
-  );
-
-  onlySessionUpdates(): acp.SessionUpdate[] {
-    return this.messages
-      .filter((m) => m.kind === 'sessionUpdate')
-      .map((m) => (m as { update: acp.SessionUpdate }).update);
-  }
-
-  sessionUpdateKinds(): string[] {
-    return this.onlySessionUpdates().map((u) => u.sessionUpdate);
-  }
-}
-
-function buildMinimalConfig(): Config {
-  return {
-    getEphemeralSetting: () => undefined,
-    getDebugMode: () => false,
-    getApprovalMode: () => 'default' as ApprovalMode,
-    setApprovalMode: () => {},
-    getTargetDir: () => '/project',
-    getFileService: () => ({ shouldIgnoreFile: () => false }),
-    getFileFilteringOptions: () => ({
-      respectGitIgnore: true,
-      respectLlxprtIgnore: true,
-    }),
-    getEnableRecursiveFileSearch: () => false,
-    getFileSystemService: () => ({ readTextFile: async () => '' }),
-    getMaxSessionTurns: () => 50,
-  } as unknown as Config;
-}
-
-function createSession(
-  agent: Agent,
-  connection: RecordingConnection,
-  config: Config = buildMinimalConfig(),
-): Session {
-  return new Session(
-    'test-session-id',
-    agent,
-    config,
-    connection as unknown as acp.AgentSideConnection,
-  );
-}
-
-async function runPrompt(session: Session): Promise<acp.PromptResponse> {
-  return session.prompt({
-    sessionId: 'test-session-id',
-    prompt: [{ type: 'text', text: 'hello' }],
-  });
-}
-
-function editConfirmation(
-  confirmationId: string,
-  toolCallId: string,
-): Extract<AgentEvent, { type: 'tool-confirmation' }> {
-  return {
-    type: 'tool-confirmation',
-    confirmation: {
-      confirmationId,
-      toolCallId,
-      name: 'edit',
-      details: {
-        type: 'edit',
-        title: 'Edit file',
-        fileName: '/project/file.txt',
-        filePath: '/project/file.txt',
-        fileDiff: 'diff',
-        originalContent: 'old',
-        newContent: 'new',
-        onConfirm: vi.fn().mockResolvedValue(undefined),
-      },
-    },
-  };
-}
+const createdSessions: Session[] = [];
 
 describe('Zed Session.prompt (Agent API) - streaming output', () => {
+  afterEach(async () => {
+    for (const session of createdSessions.splice(0)) {
+      await session.dispose();
+    }
+  });
+
   it('emits agent_message_chunk events in stream order followed by end_turn', async () => {
     const { agent } = buildFakeAgent([
       { type: 'text', text: 'Hello' },
@@ -244,6 +38,7 @@ describe('Zed Session.prompt (Agent API) - streaming output', () => {
     ]);
     const connection = new RecordingConnection();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     const response = await runPrompt(session);
 
@@ -269,6 +64,7 @@ describe('Zed Session.prompt (Agent API) - streaming output', () => {
     ]);
     const connection = new RecordingConnection();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     await runPrompt(session);
 
@@ -280,16 +76,17 @@ describe('Zed Session.prompt (Agent API) - streaming output', () => {
 
   it('preserves interleaved text and thought ordering within a batch', async () => {
     const { agent } = buildFakeAgent([
-      { type: 'text', text: 'before' },
+      { type: 'text', text: 'before. ' },
       {
         type: 'thinking',
         thought: { subject: 'thought', description: '' },
       },
-      { type: 'text', text: 'after' },
+      { type: 'text', text: 'after. ' },
       { type: 'done', reason: 'stop' },
     ]);
     const connection = new RecordingConnection();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     await runPrompt(session);
 
@@ -299,11 +96,12 @@ describe('Zed Session.prompt (Agent API) - streaming output', () => {
         text: (update as { content: { text: string } }).content.text,
       })),
     ).toStrictEqual([
-      { kind: 'agent_message_chunk', text: 'before' },
+      { kind: 'agent_message_chunk', text: 'before. ' },
       { kind: 'agent_thought_chunk', text: 'thought' },
-      { kind: 'agent_message_chunk', text: 'after' },
+      { kind: 'agent_message_chunk', text: 'after. ' },
     ]);
   });
+
   it('emits the emoji-blocked message without hanging', async () => {
     const { agent } = buildFakeAgent([
       { type: 'text', text: 'blocked \u{1F600}' },
@@ -316,6 +114,7 @@ describe('Zed Session.prompt (Agent API) - streaming output', () => {
         key === 'emojifilter' ? 'error' : undefined,
     } as unknown as Config;
     const session = createSession(agent, connection, config);
+    createdSessions.push(session);
 
     await runPrompt(session);
 
@@ -350,6 +149,7 @@ describe('Zed Session.prompt (Agent API) - tool-call status progression', () => 
     ]);
     const connection = new RecordingConnection();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     await runPrompt(session);
 
@@ -393,6 +193,7 @@ describe('Zed Session.prompt (Agent API) - tool-call status progression', () => 
     ]);
     const connection = new RecordingConnection();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     await runPrompt(session);
 
@@ -406,6 +207,42 @@ describe('Zed Session.prompt (Agent API) - tool-call status progression', () => 
       locations: [{ path: '/project/c.ts', line: 12 }],
     });
   });
+
+  it('coerces string-typed numeric line/offset args to numeric locations', async () => {
+    const { agent } = buildFakeAgent([
+      {
+        type: 'tool-call',
+        call: {
+          id: 'str-offset',
+          name: 'read_file',
+          args: { absolute_path: '/project/file.txt', offset: '7' },
+        },
+      },
+      {
+        type: 'tool-call',
+        call: {
+          id: 'str-start-line',
+          name: 'delete_line_range',
+          args: { absolute_path: '/project/c.ts', start_line: '42' },
+        },
+      },
+      { type: 'done', reason: 'stop' },
+    ]);
+    const connection = new RecordingConnection();
+    const session = createSession(agent, connection);
+    createdSessions.push(session);
+
+    await runPrompt(session);
+
+    const updates = connection.onlySessionUpdates();
+    expect(updates[0]).toMatchObject({
+      locations: [{ path: '/project/file.txt', line: 7 }],
+    });
+    expect(updates[1]).toMatchObject({
+      locations: [{ path: '/project/c.ts', line: 42 }],
+    });
+  });
+
   it('surfaces live tool-status output as tool_call_update content', async () => {
     const toolCallId = 'tool-live';
     const { agent } = buildFakeAgent([
@@ -426,6 +263,7 @@ describe('Zed Session.prompt (Agent API) - tool-call status progression', () => 
     ]);
     const connection = new RecordingConnection();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     await runPrompt(session);
 
@@ -487,6 +325,7 @@ describe('Zed Session.prompt (Agent API) - tool-call status progression', () => 
     ]);
     const connection = new RecordingConnection();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     await runPrompt(session);
 
@@ -533,6 +372,7 @@ describe('Zed Session.prompt (Agent API) - tool permission round-trip', () => {
     ]);
     const connection = new RecordingConnection();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     await runPrompt(session);
 
@@ -566,6 +406,7 @@ describe('Zed Session.prompt (Agent API) - tool permission round-trip', () => {
       optionId: ToolConfirmationOutcome.Cancel,
     });
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     const response = await runPrompt(session);
 
@@ -587,6 +428,7 @@ describe('Zed Session.prompt (Agent API) - tool permission round-trip', () => {
     const connection = new RecordingConnection();
     connection.rejectPermission(new Error('permission transport failed'));
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     await expect(runPrompt(session)).rejects.toThrow(
       /permission transport failed/,
@@ -611,6 +453,7 @@ describe('Zed Session.prompt (Agent API) - tool permission round-trip', () => {
       payload: { editedCommand: '  echo hi  ', newContent: 'replacement' },
     } as acp.RequestPermissionOutcome);
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     await runPrompt(session);
 
@@ -631,6 +474,7 @@ describe('Zed Session.prompt (Agent API) - cancellation', () => {
       buildFakeAgent([{ type: 'done', reason: 'aborted' }]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(aborted);
     await expect(runPrompt(aborted)).resolves.toStrictEqual({
       stopReason: 'cancelled',
     });
@@ -639,6 +483,7 @@ describe('Zed Session.prompt (Agent API) - cancellation', () => {
       buildFakeAgent([{ type: 'done', reason: 'max-turns' }]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(maxTurns);
     await expect(runPrompt(maxTurns)).resolves.toStrictEqual({
       stopReason: 'max_turn_requests',
     });
@@ -647,6 +492,7 @@ describe('Zed Session.prompt (Agent API) - cancellation', () => {
       buildFakeAgent([{ type: 'done', reason: 'context-overflow' }]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(contextOverflow);
     await expect(runPrompt(contextOverflow)).resolves.toStrictEqual({
       stopReason: 'max_tokens',
     });
@@ -655,6 +501,7 @@ describe('Zed Session.prompt (Agent API) - cancellation', () => {
       buildFakeAgent([{ type: 'done', reason: 'error' }]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(errorSession);
     await expect(runPrompt(errorSession)).rejects.toThrow(
       /terminal reason: error/,
     );
@@ -663,6 +510,7 @@ describe('Zed Session.prompt (Agent API) - cancellation', () => {
       buildFakeAgent([{ type: 'done', reason: 'hook-stopped' }]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(hookStoppedSession);
     await expect(runPrompt(hookStoppedSession)).rejects.toThrow(
       /terminal reason: hook-stopped/,
     );
@@ -678,6 +526,7 @@ describe('Zed Session.prompt (Agent API) - cancellation', () => {
       ]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(session);
 
     await expect(runPrompt(session)).rejects.toMatchObject({
       code: 429,
@@ -696,6 +545,7 @@ describe('Zed Session.prompt (Agent API) - cancellation', () => {
     const connection = new RecordingConnection();
     const gate = connection.armPermissionGate();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     const promptPromise = runPrompt(session);
     await gate.arrived;
@@ -736,6 +586,7 @@ describe('Zed Session.prompt (Agent API) - cancellation', () => {
     const connection = new RecordingConnection();
     const gate = connection.armPermissionGate();
     const session = createSession(agent, connection);
+    createdSessions.push(session);
 
     const firstPrompt = runPrompt(session);
     await gate.arrived;
@@ -760,6 +611,7 @@ describe('Zed Session.prompt (Agent API) - previously-dropped event variants', (
       ]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(notice);
     await expect(runPrompt(notice)).resolves.toStrictEqual({
       stopReason: 'end_turn',
     });
@@ -768,6 +620,7 @@ describe('Zed Session.prompt (Agent API) - previously-dropped event variants', (
       buildFakeAgent([{ type: 'invalid-stream' }]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(invalid);
     await expect(runPrompt(invalid)).rejects.toThrow(/invalid stream/i);
 
     const hookBlocked = createSession(
@@ -779,6 +632,7 @@ describe('Zed Session.prompt (Agent API) - previously-dropped event variants', (
       ]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(hookBlocked);
     await expect(runPrompt(hookBlocked)).rejects.toThrow(
       /Blocked by pre-tool hook/,
     );
@@ -802,6 +656,7 @@ describe('Zed Session.prompt (Agent API) - previously-dropped event variants', (
       ]).agent,
       new RecordingConnection(),
     );
+    createdSessions.push(ignored);
     await expect(runPrompt(ignored)).resolves.toStrictEqual({
       stopReason: 'end_turn',
     });
